Lock Down Mutations

+++ src/resolvers/Mutation.js
@@ -1,35 +1,70 @@
 import bcrypt from 'bcryptjs'
+import jwt from 'jsonwebtoken'
+import getUserID from "../utils/getUserID"
 
 const Mutation = {
+  // 
+  // actions
+  // 
+  async login(parent, args, { prisma }, info) {
+    const user = await prisma.query.user({
+      where: {
+        email: args.data.email
+      }
+    })
+    if (!user) {
+      throw new Error('Unable to login')
+    }
+    const isMatch = await bcrypt.compare(args.data.password, user.password)
+    if (!isMatch) {
+      throw new Error('Unable to login')
+    }
+    return {
+      user,
+      token: jwt.sign({ userId: user.id }, "popsicle")
+    }
+  },
+  // 
+  // USER 
+  // 
   async createUser(parent, args, { prisma }, info) {
     if (args.data.password.length < 8) {
       throw new Error('Password must be 8 characters or longer.')
     }
     const password = await bcrypt.hash(args.data.password, 10)
-
-    return await prisma.mutation.createUser({
+    const user = await prisma.mutation.createUser({
       data: {
         ...args.data,
         password
       }
-    }, info)
+    })
+    return {
+      user,
+      token: jwt.sign({ userId: user.id }, "popsicle")
+    }
   },
-  deleteUser(parent, args, { prisma }, info) {
-    return prisma.mutation.deleteUser({
-      where: {
-        id: args.id
-      }
-    }, info)
-  },
-  async updateUser(parent, args, { prisma }, info) {
+  async updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserID(request)
     return prisma.mutation.updateUser({
       where: {
-        id: args.id
+        id: userId
       },
       data: args.data
     }, info)
   },
-  createPost(parent, args, { prisma }, info) {
+  deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserID(request)
+    return prisma.mutation.deleteUser({
+      where: {
+        id: userId
+      }
+    }, info)
+  },
+  // 
+  // Post 
+  // 
+  async  createPost(parent, args, { prisma, request }, info) {
+    const userId = getUserID(request)
     return prisma.mutation.createPost({
       data: {
         title: args.data.title,
@@ -37,20 +72,26 @@ const Mutation = {
         published: args.data.published,
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         }
       }
     }, info)
   },
-  deletePost(parent, args, { prisma }, info) {
-    return prisma.mutation.deletePost({
-      where: {
-        id: args.id
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserID(request)
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
       }
-    }, info)
-  },
-  updatePost(parent, args, { prisma }, info) {
+    })
+
+
+    if (!postExists) {
+      throw new Error('Unable to update post')
+    }
+
     return prisma.mutation.updatePost({
       where: {
         id: args.id
@@ -58,13 +99,38 @@ const Mutation = {
       data: args.data
     }, info)
   },
-  createComment(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserID(request)
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!postExists) {
+      throw new Error('Unable to delete post')
+    }
+
+    return prisma.mutation.deletePost({
+      where: {
+        id: args.id
+      }
+    }, info)
+  },
+
+  // 
+  // Comment 
+  // 
+  async createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserID(request)
+
     return prisma.mutation.createComment({
       data: {
         text: args.data.text,
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         },
         post: {
@@ -75,19 +141,45 @@ const Mutation = {
       }
     }, info)
   },
-  deleteComment(parent, args, { prisma }, info) {
-    return prisma.mutation.deleteComment({
-      where: {
-        id: args.id
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserID(request)
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
       }
-    }, info)
-  },
-  updateComment(parent, args, { prisma }, info) {
+    })
+
+    if (!commentExists) {
+      throw new Error('Unable to delet comment')
+    }
+
     return prisma.mutation.updateComment({
       where: {
-        id: args.id
+        id: userId
       },
       data: args.data
+    }, info)
+  },
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserID(request)
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!commentExists) {
+      throw new Error('Unable to delet comment')
+    }
+
+    return prisma.mutation.deleteComment({
+      where: {
+        id: userId
+      }
     }, info)
   }
 }
